+++ bin/pyre_parse.ml
@@ -9,8 +9,9 @@ let read_file_content = function
 let parse ~f content =
   Parser.with_context (fun context ->
       f ~context content
-      |> Result.map_error (fun { Parser.Error.message; line; column } ->
-             Format.sprintf "Parse error at line %d, column %d: %s" line column message))
+      |> Result.map_error (fun { Parser.Error.message; line; column; end_line; end_column } ->
+             Format.sprintf "Parse error at line %d, column %d to line %d, column %d: %s" line
+               column end_line end_column message))
 
 let handle_result ~f = function
   | Result.Ok result -> f result