+++ test/misc.t
@@ -1,8 +1,8 @@
 Test parse error
-  $ echo "def foo(" | parse module -
-  Parse error at line 1, column 8: '(' was never closed
+  $ echo "def foo(" | parse module -  # The funny end column 0 came from CPython
+  Parse error at line 1, column 8 to line 1, column 0: '(' was never closed
   $ echo "del 1+2" | parse module -
-  Parse error at line 1, column 5: cannot delete expression
+  Parse error at line 1, column 5 to line 1, column 8: cannot delete expression
 
 Non-ascii characters in string literals is ok
   $ echo 'x = "😝"' | parse module -
@@ -24,7 +24,7 @@ Non-ascii characters in string literals is ok
 
 Non-ascii characters in variable name is not ok
   $ echo 'Δx' | parse module -
-  Parse error at line 1, column 0: CPython runtime raised a non-syntax exception
+  Parse error at line 1, column 1 to line 1, column 1: CPython runtime raised a non-syntax exception
 
 Unicode surrogate should not crash the parser
   $ echo '"\ud83d"' | parse module -